+++ src/process/process.h
@@ -1,4 +1,7 @@
 #pragma once
+#include <stddef.h>
+#include <stdint.h>
+
 #include "../buffer/synchronous_single_buffer.h"
 
 typedef struct {
@@ -12,5 +15,5 @@ typedef struct {
 } AudioFeatures;
 
 void* reader_thread(const ReaderContext* rc);
-double get_rms(const int *arr, int num_of_elements);
-double get_zcr(const int *arr, int num_of_elements);
+double get_rms(const int16_t* arr, size_t num_of_elements);
+double get_zcr(const int16_t* arr, size_t num_of_elements);
